BUGFIX and REVERSION
somewhere along the way in development of the binary branch the MI
calucaltions went wonky. To correct this I reverted to an old good
master branch, double checked that, then manually added back the key
bits of the binary branch.

This commit is totally incompatible with priors on this branch, but is
semi compatible with master, for now anyway.